A CI step to check that gazelle and buildifier has been run (#32)
* Add a check for unchanged files

* add buildifier and fix lint errors

* commit buildifier fixes

* gazelle fixes and rework the project structure

Usually the command applications are kept in a folder called `cmd`. This
is especially relevant if there are multiple CLI applications. Hence the
move of the CLI library code to the `internal` package.

What is more, the `main` package is normally reserved for executable
applications. In our case the top level docs package may still be called
`zserio`.

* fix remaining buildifier issues