Suppoer running tests without bazel
Running the tests require a bit of preparation: generate test data using
the Python zserio implementation, and generating Go code for the test
zserio schema. Bazel normally handles all of that automatically, but to
support people who do not want to use bazel there is now a
prepare-testdata script that will do the same thing.