Use interfaces instead of structs in Marshaling interfaces
After the introduction of the new writer struct, it seems that there is
no performance penalty to use interfaces in the interface definitions
and the generated code. This commit is doing a minimum amount of things
in order to keep things tidy and be easy to revert at the same time.