Remove replace statements from go.mod
We used a replace statement to force use of a recent version of
github.com/miekg/dns which does not suffer from CVE-2017-15133,
Unfortunately this resulted in go-zserio no longer being installable:

  $ go install github.com/woven-planet/go-zserio/cmd/go-zserio@latest
  go: github.com/woven-planet/go-zserio/cmd/go-zserio@latest (in github.com/woven-planet/go-zserio@v0.2.6):
      The go.mod file for the module providing named packages contains one or
      more replace directives. It must not contain directives that would cause
      it to be interpreted differently than if it were the main module.

There is a long discussion surrounding this topic at golang/go#44840,
but the TLDR is that that was a deliberate design choice and is not
likely to change anytime soon.

Removing the replace statements means an insecure version of miegk/dns
is picked, but since that is not used in go-zserio we can ignore the
security warning.