Fix Error when evaluating subtyped Enums
This issue was reported here: https://github.com/woven-planet/go-zserio/issues/108

To solve this issue, it is necessary to fully look up the original
type of an expression, to remove all indirections caused by subtyping.
Types such as enums that get type definition using the "subtype" type,
require resolving the original type that was used.
This code addresses this issue as follows:
1) If case of a dot expression, the code now looks up the original type,
   to remove subtype indirections.
2) Renamed / refactored the original type lookup a little bit. There is
   quite a bit of code duplication that should be further refactored
   (future work).
3) Renamed functions for better wording.