Add a possibility to lookup the original type from the scope
- This is needed in cases the original (not subtyped) type is needed.
- An example is Enum value generation:
Due to Go not supporting Enums, we need to check if subtyping is used.
instead of generating the Enum value with the subtyped name, we need the
original type name.
For example, instead of "SubTypeColorBlue", we want "ColorBlue",
because only the latter one is defined in the generated Go code.

This makes it necessary to look up the original zserio type during code
generation.